maxbalance config, storage alert
